requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.36
aiosqlite==0.19.0
asyncpg>=0.29.0

# Pydantic and validation
pydantic>=2.10.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# AI / LLM (Ollama for local, Groq for cloud)
ollama>=0.4.0
groq>=0.4.0

# Authentication
passlib[bcrypt]==1.7.4
bcrypt<4.1
python-jose[cryptography]==3.3.0

# Env
python-dotenv==1.0.0
